\name{xMWAS_0.54-package}
\alias{xMWAS_0.54-package}
\alias{xMWAS}
\docType{package}
\title{xMWAS}
\description{
xMWAS allows integration of omics data using multivariate methods such as PLS. The package performs community detection and network analysis to allow visualization of positive or negative associations between different datasets generated using samples from same individuals. 
The run_xmwas() function performs the integrative, community detection, and network analysis. The output files include .gml files that can be used with software like Cytoscape.
}
\details{
\tabular{ll}{
Package: \tab xMWAS\cr
Type: \tab Package\cr
Version: \tab 0.54\cr
Date: \tab 2018-03-01\cr
License: \tab gpl2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Karan Uppal
Maintainer: <kuppal2@emory.edu>
}
\keyword{ xMWAS }

